Consolidate CharInv pieces
