+++ src/goals/CharacterInventory/Redux/CharacterInventoryActions.ts
@@ -1,7 +1,7 @@
-import { Action, PayloadAction } from "@reduxjs/toolkit";
+import { type Action, type PayloadAction } from "@reduxjs/toolkit";
 
-import { Project } from "api/models";
-import { getFrontierWords } from "backend";
+import { type Project } from "api/models";
+import { getFrontierWords, updateWord } from "backend";
 import router from "browserRouter";
 import {
   addCharInvChangesToGoal,
@@ -9,8 +9,8 @@ import {
 } from "components/GoalTimeline/Redux/GoalActions";
 import { asyncUpdateCurrentProject } from "components/Project/ProjectActions";
 import {
+  type CharacterChange,
   CharacterStatus,
-  CharacterChange,
 } from "goals/CharacterInventory/CharacterInventoryTypes";
 import {
   addRejectedCharacterAction,
@@ -23,12 +23,12 @@ import {
   setValidCharactersAction,
 } from "goals/CharacterInventory/Redux/CharacterInventoryReducer";
 import {
-  CharacterInventoryState,
-  CharacterSetEntry,
+  type CharacterInventoryState,
+  type CharacterSetEntry,
   getCharacterStatus,
 } from "goals/CharacterInventory/Redux/CharacterInventoryReduxTypes";
-import { StoreState } from "types";
-import { StoreStateDispatch } from "types/Redux/actions";
+import { type StoreState } from "types";
+import { type StoreStateDispatch } from "types/Redux/actions";
 import { Path } from "types/path";
 
 // Action Creation Functions
@@ -69,6 +69,8 @@ export function setValidCharacters(chars: string[]): PayloadAction {
 
 // Dispatch Functions
 
+/** Returns a dispatch function to: update the in-state `.validCharacters` and
+ * `.rejectedCharacters` according to the given character and status. */
 export function setCharacterStatus(character: string, status: CharacterStatus) {
   return (dispatch: StoreStateDispatch, getState: () => StoreState) => {
     switch (status) {
@@ -97,7 +99,7 @@ export function setCharacterStatus(character: string, status: CharacterStatus) {
   };
 }
 
-/** Sends the in-state character inventory to the server. */
+/** Returns a dispatch function to: send the in-state char inventory to the server. */
 export function uploadInventory() {
   return async (dispatch: StoreStateDispatch, getState: () => StoreState) => {
     const charInvState = getState().characterInventoryState;
@@ -120,6 +122,8 @@ export function uploadInventory() {
   };
 }
 
+/** Returns a dispatch function to: fetch the current project's frontier and, from those
+ * words, update the array of all vernacular forms in-state. */
 export function fetchWords() {
   return async (dispatch: StoreStateDispatch) => {
     const words = await getFrontierWords();
@@ -127,6 +131,9 @@ export function fetchWords() {
   };
 }
 
+/** Returns a dispatch function to: gather all characters (and number of occurrences)
+ * in the in-state `allWords` array and extract the inventory status of each character
+ * from the current project's `.validCharacters` and `.rejectedCharacters`. */
 export function getAllCharacters() {
   return async (dispatch: StoreStateDispatch, getState: () => StoreState) => {
     const allWords = getState().characterInventoryState.allWords;
@@ -143,6 +150,8 @@ export function getAllCharacters() {
   };
 }
 
+/** Returns a dispatch function to: load all character inventory data for the current
+ * project (drawing from the in-state project and its frontier in the database.) */
 export function loadCharInvData() {
   return async (dispatch: StoreStateDispatch, getState: () => StoreState) => {
     const project = getState().currentProjectState.project;
@@ -154,12 +163,37 @@ export function loadCharInvData() {
   };
 }
 
+/** Returns a dispatch function to: update every word in the current project's frontier
+ * that has the given `findValue` in its vernacular form. Then:
+ * - Update the in-state `allWords` array;
+ * - Update the in-state character inventory. */
+export function findAndReplace(findValue: string, replaceValue: string) {
+  return async (dispatch: StoreStateDispatch) => {
+    const changedWords = (await getFrontierWords()).filter((w) =>
+      w.vernacular.includes(findValue)
+    );
+    const findRegExp = new RegExp(
+      findValue.replace(/[-\\^$*+?.()|[\]{}]/g, "\\$&"),
+      "g"
+    );
+    for (const word of changedWords) {
+      word.vernacular = word.vernacular.replace(findRegExp, replaceValue);
+      await updateWord(word);
+    }
+    await dispatch(fetchWords());
+    await dispatch(getAllCharacters());
+  };
+}
+
 // Helper Functions
 
+/** Navigate to the Data Cleanup page. */
 export function exit(): void {
   router.navigate(Path.Goals);
 }
 
+/** Count the number of occurrences of the given `char` in the given array of strings.
+ * Gives a console error if `char` is not length 1. */
 function countOccurrences(char: string, words: string[]): number {
   if (char.length !== 1) {
     console.error(`countOccurrences expects length 1 char, but got: ${char}`);
@@ -175,6 +209,8 @@ function countOccurrences(char: string, words: string[]): number {
   return count;
 }
 
+/** Compare the `.validCharacters` and `.rejectedCharacters` between the given project
+ * and the given state to identify all characters with a change of inventory status. */
 export function getChanges(
   project: Project,
   charInvState: CharacterInventoryState
@@ -196,7 +232,8 @@ export function getChanges(
   return changes;
 }
 
-// Returns undefined if CharacterStatus unchanged.
+/** Return the given character's change of inventory status, or undefined if
+ * its CharacterStatus is unchanged. */
 function getChange(
   c: string,
   oldAcc: string[],
